remove state from App.js connect react with redux in GroceryItems component
