+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/App.js":"1","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryCart.js":"2","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryItems.js":"3","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/index.js":"4","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/reducers/cart-reducer.js":"5"},{"size":914,"mtime":1610965982112,"results":"6","hashOfConfig":"7"},{"size":1145,"mtime":1610965740636,"results":"8","hashOfConfig":"7"},{"size":1486,"mtime":1610964363663,"results":"9","hashOfConfig":"7"},{"size":448,"mtime":1610966934323,"results":"10","hashOfConfig":"7"},{"size":273,"mtime":1610967236443,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"9hsum9",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/App.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryCart.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryItems.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/index.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/reducers/cart-reducer.js",["22","23"],{"ruleId":"24","severity":1,"message":"25","line":12,"column":5,"nodeType":"26","messageId":"27","endLine":14,"endColumn":6},{"ruleId":"24","severity":1,"message":"28","line":16,"column":5,"nodeType":"26","messageId":"29","endLine":17,"endColumn":20},"no-fallthrough","Expected a 'break' statement before 'case'.","SwitchCase","case","Expected a 'break' statement before 'default'.","default"]
+[{"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/App.js":"1","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryCart.js":"2","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryItems.js":"3","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/index.js":"4","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/reducers/cart-reducer.js":"5"},{"size":383,"mtime":1610968971460,"results":"6","hashOfConfig":"7"},{"size":1145,"mtime":1610965740636,"results":"8","hashOfConfig":"7"},{"size":1176,"mtime":1610968730901,"results":"9","hashOfConfig":"7"},{"size":448,"mtime":1610966934323,"results":"10","hashOfConfig":"7"},{"size":951,"mtime":1610968559169,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"9hsum9",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/App.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryCart.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/components/GroceryItems.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/index.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-cashier/src/reducers/cart-reducer.js",["22","23"],{"ruleId":"24","severity":1,"message":"25","line":31,"column":5,"nodeType":"26","messageId":"27","endLine":33,"endColumn":6},{"ruleId":"24","severity":1,"message":"28","line":35,"column":5,"nodeType":"26","messageId":"29","endLine":36,"endColumn":20},"no-fallthrough","Expected a 'break' statement before 'case'.","SwitchCase","case","Expected a 'break' statement before 'default'.","default"]
