+++ src/components/GroceryItems.js
@@ -1,24 +1,7 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
 
-const GROCERY_ITEMS = [
-  { name: "Sliced bacon", price: 5.79 },
-  { name: "Pasta", price: 1.28 },
-  { name: "Dried beans", price: 1.36 },
-  { name: "Ground beef", price: 4.12 },
-  { name: "All-purpose flour", price: 0.52 },
-  { name: "Creamy peanut butter", price: 2.56 },
-  { name: "Top round steak", price: 5.78 },
-  { name: "Potatoes", price: 0.72 },
-  { name: "Frozen turkey", price: 1.59 },
-  { name: "Sirloin steak", price: 8.07 },
-  { name: "White rice", price: 0.72 },
-  { name: "Chocolate chip cookies", price: 3.47 },
-  { name: "Seedless grapes", price: 2.67 },
-  { name: "Sugar", price: 0.65 },
-  { name: "Ice cream", price: 4.70 },
-]
-
-export default class GroceryItems extends Component {
+class GroceryItems extends Component {
   render () {
     return(
       <div id="grocery-items">
@@ -31,7 +14,7 @@ export default class GroceryItems extends Component {
             </tr>
           </thead>
           <tbody>
-            {GROCERY_ITEMS.map((item, index) => {
+            {this.props.items.map((item, index) => {
               return(
                 <tr key={index}>
                   <td>
@@ -50,3 +33,21 @@ export default class GroceryItems extends Component {
     )
   }
 }
+
+function mapStateToProps(state) {
+  return {
+    items: state.forSale
+  }
+}
+
+function mapDispatchToProps(dispatch) {
+  return {
+    addToCart: (item) => {
+      dispatch({
+        type: 'ADD_TO_CART',
+        item
+      })
+    }
+  }
+}
+export default connect(mapStateToProps, mapDispatchToProps)(GroceryItems);
